Update seraph.js
Update of _createNodeObject to address overwriting of ['id'] property of the object. Currently if the object in the database has property named ID that is not the internal ID assigned to the node, the function overwrites it with internal ID. This quick fix checks to see if ID already exists, if it doesn't it allows the internal ID to be used.